# api/openapi.yaml
openapi: 3.1.0
info:
  title: Jewelcase API
  version: 1.0.0-draft
  summary: The public API of a Jewelcase server, used by first-party and third-party clients alike.
  description: |
    The contract between Jewelcase servers and every client. First-party clients use
    nothing outside it (`design/general.md` §6). Requirements cited as
    `requirements/file.md` §N.

    ## Conventions

    **Base path.** Paths are relative to `{basePath}/api/v1`. A server may sit under any
    prefix behind a reverse proxy; clients never assume the domain root.

    **Authentication.** Login registers a device and returns an opaque token. Send it as
    `Authorization: Bearer <token>`. The web client may instead rely on the
    `jewelcase_session` cookie set at login; cookie-authenticated requests that change
    state must carry an `X-Jewelcase-Client` header or they are refused. Any authenticated
    endpoint may answer `401` — see `Problem.reason` for whether a device keeps its
    downloads (`requirements/offline.md` §9).

    **Scope.** Paths under `/libraries/{libraryId}` are scoped to that library. Anything
    the caller cannot reach answers `404`, exactly as if it did not exist; an action the
    caller's role forbids answers `403` (`requirements/users.md` §10). Identifiers in a
    request select within the caller's scope and never widen it.

    **Personal data.** `/me` is the caller's own account. Admins read other accounts'
...

...

...
    never on a collection. Lists report `total`.

...

...
servers:
...
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /setup:
    post:
      tags: [Auth]
      summary: Create the owner account
      description: |
        First-run setup (`requirements/users.md` §2.1). Until it completes, every other
        endpoint except `/health` and `/server` answers `503 setup_required`. Once an owner
        exists this endpoint answers `404`.
      operationId: completeSetup
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetupRequest' }
      responses:
        '201':
          description: Owner created and logged in on this device.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        default: { $ref: '#/components/responses/Problem' }
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in and register a device
      description: |
        Failure is identical for an unknown account, a wrong password, and a suspended
        account (`requirements/users.md` §3). Repeated failures are delayed, then
        temporarily locked, per account and per origin (`requirements/users.md` §3.1).
        A device logged out with `credentials_changed` passes its `deviceId` to log back in
        as itself.
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Logged in. Also sets the session cookie for browser clients.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
        '401':
          description: '`invalid_credentials`, whatever the cause.'
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
        default: { $ref: '#/components/responses/Problem' }
  /auth/logout:
    post:
      tags: [Auth]
      summary: Log out this device
      description: Removes the device. Its downloads are removed (`requirements/offline.md` §9).
      operationId: logout
      responses:
        '204': { description: Logged out. }
        default: { $ref: '#/components/responses/Problem' }
...
  /me:
    get:
      tags: [Account]
      summary: Get the caller's account
      operationId: getMe
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        default: { $ref: '#/components/responses/Problem' }
    patch:
      tags: [Account]
      summary: Update profile
      operationId: updateMe
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserUpdate' }
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        default: { $ref: '#/components/responses/Problem' }
...
  /me/plugins:
    get:
      tags: [Account]
      summary: List the plugins the caller can connect
      description: |
        Plugins that act for each person, such as a scrobbler, connected with the person's own
        account on its service (`requirements/plugins.md` §4.1, §6): every installed plugin
        with personal settings that is enabled in a library the caller reaches.
      operationId: listMyPlugins
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/PersonalPlugin' }
        default: { $ref: '#/components/responses/Problem' }
  /me/plugins/{pluginId}/settings:
    parameters:
      - $ref: '#/components/parameters/PluginId'
    get:
      tags: [Account]
      summary: Get the caller's settings for a plugin
      description: |
        Per-user settings and credentials for personal sources (`requirements/plugins.md` §6),
        as the plugin's `personalSettings` declare them. They are the person's, the same in
        every library. Secrets are never returned. Any plugin not in `listMyPlugins` answers
        `404`.
      operationId: getMyPluginSettings
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PluginSettings' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Problem' }
    put:
      tags: [Account]
      summary: Connect a plugin, or change what it is connected with
      description: |
        Every required setting must have a value. Validated by the plugin when entered, run with
        these and the server-wide settings; an invalid value answers
        `422 plugin_settings_invalid`. The plugin receives the caller's listening from then on,
        never what came before.
      operationId: setMyPluginSettings
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PluginSettingsUpdate' }
      responses:
        '200':
          description: Saved.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PluginSettings' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        default: { $ref: '#/components/responses/Problem' }
    delete:
      tags: [Account]
      summary: Disconnect a plugin
      description: |
        Forgets the caller's settings for it; it stops acting for them. Sharing their searches
        with it stops too, as `setMySearchSharing` turning it off does.
      operationId: disconnectMyPlugin
      responses:
        '204': { description: Disconnected, or never connected. }
        default: { $ref: '#/components/responses/Problem' }
  /me/plugins/{pluginId}/search-sharing:
    parameters:
      - $ref: '#/components/parameters/PluginId'
    put:
      tags: [Account]
      summary: Turn sharing the caller's searches with a plugin on or off
      description: |
        Off until the caller turns it on (`requirements/search.md` §6). On, the plugin is sent
        each search the caller settles on from then, never one before. Off, or disconnecting,
        asks it to forget every search it was sent; what it already sent on is not recalled
        (`requirements/users.md` §7). Turning it on needs the caller to have connected the
        plugin, and the plugin to ask for searches (`asksForSearches`), or answers `422`.
      operationId: setMySearchSharing
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SearchSharing' }
      responses:
        '204': { description: Set. }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        default: { $ref: '#/components/responses/Problem' }

  # ───────────────────────────── Devices ─────────────────────────────
...
  /libraries:
    get:
      tags: [Libraries]
      summary: List libraries the caller can access
      description: Empty for an account with no access; clients explain why rather than erroring (`requirements/users.md` §5).
      operationId: listLibraries
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Library' }
        default: { $ref: '#/components/responses/Problem' }
  /libraries/{libraryId}:
    parameters:
      - $ref: '#/components/parameters/LibraryId'
    get:
      tags: [Libraries]
      summary: Get a library
      operationId: getLibrary
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Library' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Problem' }

  # ───────────────────────────── Browse ─────────────────────────────
  /libraries/{libraryId}/tracks:
    parameters:
      - $ref: '#/components/parameters/LibraryId'
    get:
      tags: [Browse]
      summary: Browse tracks
      description: |
        Any filtered list of the library's tracks. Filters compose (`requirements/tags.md` §7).
        An album's track listing is `albumId` with `sort=album`: disc then track order, missing
        tracks in position (`requirements/albums.md` §4). Whether a track is owned or featured
        for an artist follows from its `album.artists`. Playlists, mixes, and queues are ordered
        item lists, not track filters, and have their own item endpoints.
      operationId: listTracks
      parameters:
        - name: sort
          in: query
          description: |
            `album` orders by album, then disc, then track number. `top` is the caller's own
            ranking — listening weighted by completion, not play count alone — falling back to a
            sensible order for tracks never played, so it is never empty
            (`requirements/artists.md` §3.3, `requirements/analytics.md` §4). An artist's top
            tracks are `artistId` with `sort=top`.
          schema:
            type: string
            enum: [name, artist, album, dateAdded, playCount, lastPlayed, releaseDate, top]
            default: name
        - $ref: '#/components/parameters/Order'
        - name: albumId
          in: query
          description: Only tracks on this album.
          schema: { type: string }
        - $ref: '#/components/parameters/ArtistFilter'
        - $ref: '#/components/parameters/ArtistCredit'
        - $ref: '#/components/parameters/Genre'
        - $ref: '#/components/parameters/GenreMatch'
        - $ref: '#/components/parameters/YearFrom'
        - $ref: '#/components/parameters/YearTo'
        - $ref: '#/components/parameters/Availability'
        - $ref: '#/components/parameters/Played'
        - $ref: '#/components/parameters/PlayedBefore'
        - $ref: '#/components/parameters/PlayedAfter'
        - $ref: '#/components/parameters/MinPlayCount'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TrackPage' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Problem' }
  /libraries/{libraryId}/tracks/{trackId}:
    parameters:
      - $ref: '#/components/parameters/LibraryId'
      - $ref: '#/components/parameters/TrackId'
    get:
      tags: [Browse]
      summary: Get a track
      operationId: getTrack
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Track' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Problem' }
  /libraries/{libraryId}/albums:
    parameters:
      - $ref: '#/components/parameters/LibraryId'
    get:
      tags: [Browse]
      summary: Browse albums
      description: |
        Any filtered list of the library's albums. An artist's discography is `artistId` with
        `artistCredit=owned`, grouped by `type` with empty sections omitted, newest first by
        default (`requirements/artists.md` §3.1). Their appearances are `artistCredit=featured`;
        the tracks on each come from `listTracks` with `artistId`, `artistCredit=featured`, and
        `sort=album` (`requirements/artists.md` §3.2).
      operationId: listAlbums
      parameters:
        - name: sort
          in: query
          schema:
            type: string
            enum: [name, artist, dateAdded, playCount, lastPlayed, releaseDate, duration]
            default: name
        - $ref: '#/components/parameters/Order'
        - $ref: '#/components/parameters/ArtistFilter'
        - $ref: '#/components/parameters/ArtistCredit'
        - name: type
          in: query
          description: Filter by album type (`requirements/albums.md` §6).
          style: form
          explode: false
          schema:
            type: array
            items: { $ref: '#/components/schemas/AlbumType' }
        - $ref: '#/components/parameters/Genre'
        - $ref: '#/components/parameters/GenreMatch'
        - $ref: '#/components/parameters/YearFrom'
        - $ref: '#/components/parameters/YearTo'
        - $ref: '#/components/parameters/Availability'
        - $ref: '#/components/parameters/Played'
        - $ref: '#/components/parameters/PlayedBefore'
        - $ref: '#/components/parameters/PlayedAfter'
        - $ref: '#/components/parameters/MinPlayCount'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AlbumPage' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Problem' }
  /libraries/{libraryId}/albums/{albumId}:
    parameters:
      - $ref: '#/components/parameters/LibraryId'
      - $ref: '#/components/parameters/AlbumId'
    get:
      tags: [Browse]
      summary: Get an album
      description: |
        The album with a summary of each disc present. Its tracks are listed with `listTracks`
        using `albumId` and `sort=album`, since an album's size is unbounded — the unknown album
        holds every file with no album tag.
      operationId: getAlbum
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Album' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Problem' }
  /libraries/{libraryId}/artists:
    parameters:
      - $ref: '#/components/parameters/LibraryId'
    get:
      tags: [Browse]
      summary: Browse artists
      description: Includes artists credited only as features (`requirements/artists.md` §5).
      operationId: listArtists
      parameters:
        - name: sort
          in: query
          schema:
            type: string
            enum: [name, dateAdded, playCount, lastPlayed]
            default: name
        - $ref: '#/components/parameters/Order'
        - $ref: '#/components/parameters/Genre'
        - $ref: '#/components/parameters/GenreMatch'
        - $ref: '#/components/parameters/Availability'
        - $ref: '#/components/parameters/Played'
        - $ref: '#/components/parameters/PlayedBefore'
        - $ref: '#/components/parameters/PlayedAfter'
        - $ref: '#/components/parameters/MinPlayCount'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ArtistPage' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Problem' }
  /libraries/{libraryId}/artists/{artistId}:
    parameters:
      - $ref: '#/components/parameters/LibraryId'
      - $ref: '#/components/parameters/ArtistId'
    get:
      tags: [Browse]
      summary: Get an artist
      operationId: getArtist
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Artist' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Problem' }
...
  /libraries/{libraryId}/images/{imageId}:
    parameters:
      - $ref: '#/components/parameters/LibraryId'
      - name: imageId
        in: path
        required: true
        schema: { type: string }
        description: Content-addressed; replacing an image on disk yields a new ID.
    get:
      tags: [Media]
      summary: Get an image
      description: |
        Album, artist, and playlist images, resized to the requested size and
        format-negotiated by `Accept` (`requirements/conventions.md` §5). Immutable and
        cacheable indefinitely, because its ID changes whenever the image does.
      operationId: getImage
      parameters:
        - $ref: '#/components/parameters/ImageSize'
      responses:
        '200': { $ref: '#/components/responses/Image' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Problem' }
  /libraries/{libraryId}/tracks/{trackId}/lyrics:
    parameters:
      - $ref: '#/components/parameters/LibraryId'
      - $ref: '#/components/parameters/TrackId'
    get:
      tags: [Media]
      summary: Get lyrics
      description: A track without lyrics answers `200` with `kind` `none` (`requirements/tracks.md` §5).
      operationId: getLyrics
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Lyrics' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Problem' }
  /libraries/{libraryId}/tracks/{trackId}/waveform:
    parameters:
      - $ref: '#/components/parameters/LibraryId'
      - $ref: '#/components/parameters/TrackId'
    get:
      tags: [Media]
      summary: Get a waveform
      description: |
        Fetched on demand for the track playing and with downloads, never synced wholesale
        (`requirements/playback.md` §3). Deterministic for the same audio, so cacheable until the
        track's audio changes.
      operationId: getWaveform
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Waveform' }
        '204':
          description: |
            Not yet analyzed. Clients show a plain progress bar, and fetch again once the track's
            `loudness` becomes non-null.
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Problem' }
  /libraries/{libraryId}/tracks/{trackId}/playback:
    parameters:
      - $ref: '#/components/parameters/LibraryId'
      - $ref: '#/components/parameters/TrackId'
    get:
      tags: [Media]
      summary: Negotiate audio delivery
      description: |
        Chooses direct or transcoded delivery from the requested quality and what the client
        can decode: the original whenever the client can play it at or below `quality`
        (`requirements/playback.md` §1). The client picks `quality` from its own settings for
        the connection it is on, or for downloads. Describes exactly what the listener will hear.
      operationId: getPlaybackInfo
      x-status: draft
      parameters:
        - $ref: '#/components/parameters/Purpose'
        - $ref: '#/components/parameters/QualityParam'
        - $ref: '#/components/parameters/Codecs'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlaybackInfo' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Problem' }
  /libraries/{libraryId}/tracks/{trackId}/audio:
    parameters:
      - $ref: '#/components/parameters/LibraryId'
      - $ref: '#/components/parameters/TrackId'
    get:
      tags: [Media]
      summary: Stream or download audio
      description: |
        Serves a variant chosen by `getPlaybackInfo`, with byte ranges for seeking and
        resumable downloads. A new stream on a saturated host is refused with
        `503 capacity_exceeded`; streams already playing are never degraded
        (`requirements/performance.md` §7).
      operationId: getAudio
      x-status: draft
      parameters:
        - name: variant
          in: query
          required: true
          schema: { type: string }
        - name: Range
          in: header
          schema: { type: string }
      responses:
        '200': { $ref: '#/components/responses/Audio' }
        '206': { $ref: '#/components/responses/Audio' }
        '404': { $ref: '#/components/responses/NotFound' }
        '416': { description: Range not satisfiable. }
        '503': { $ref: '#/components/responses/CapacityExceeded' }
        default: { $ref: '#/components/responses/Problem' }

  # ───────────────────────────── Search ─────────────────────────────
  /libraries/{libraryId}/search:
    parameters:
      - $ref: '#/components/parameters/LibraryId'
    get:
      tags: [Search]
      summary: Search a library
      description: |
        Names and lyrics, with typo tolerance, diacritic and case folding, word-order
        independence, and exact-match precedence (`requirements/search.md` §3). Results
        are identical to on-device search over the same corpus (`requirements/search.md` §2.1).
        External sources are searched separately so they never delay these results.
      operationId: search
      parameters:
        - $ref: '#/components/parameters/Query'
        - $ref: '#/components/parameters/SearchTypes'
        - $ref: '#/components/parameters/Within'
        - $ref: '#/components/parameters/Genre'
        - $ref: '#/components/parameters/YearFrom'
        - $ref: '#/components/parameters/YearTo'
        - $ref: '#/components/parameters/Played'
        - $ref: '#/components/parameters/PlayedBefore'
        - $ref: '#/components/parameters/PlayedAfter'
        - $ref: '#/components/parameters/MinPlayCount'
        - name: sectionLimit
          in: query
          description: Results shown per section before expanding.
          schema: { type: integer, minimum: 1, maximum: 50, default: 5 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SearchResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Problem' }
  /libraries/{libraryId}/search/{section}:
    parameters:
      - $ref: '#/components/parameters/LibraryId'
      - name: section
        in: path
        required: true
        schema: { $ref: '#/components/schemas/SearchSectionType' }
    get:
      tags: [Search]
      summary: Expand a search section
      description: The full list for one section, in match order or sorted (`requirements/search.md` §4).
      operationId: searchSection
      parameters:
        - $ref: '#/components/parameters/Query'
        - $ref: '#/components/parameters/Within'
        - $ref: '#/components/parameters/Genre'
        - $ref: '#/components/parameters/YearFrom'
        - $ref: '#/components/parameters/YearTo'
        - $ref: '#/components/parameters/Played'
        - $ref: '#/components/parameters/PlayedBefore'
        - $ref: '#/components/parameters/PlayedAfter'
        - $ref: '#/components/parameters/MinPlayCount'
        - name: sort
          in: query
          description: '`relevance`, or any sort the section''s browse endpoint accepts.'
          schema: { type: string, default: relevance }
        - $ref: '#/components/parameters/Order'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SearchResultPage' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Problem' }
...
  /libraries/{libraryId}/recent-searches/{recentSearchId}:
    parameters:
      - $ref: '#/components/parameters/LibraryId'
      - name: recentSearchId
        in: path
        required: true
        schema: { type: string }
    delete:
      tags: [Search]
      summary: Remove a recent search
      operationId: deleteRecentSearch
      responses:
        '204': { description: Removed. }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Problem' }

  # ───────────────────────────── Sync ─────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Opaque per-device token from login.
    cookieAuth:
      type: apiKey
      in: cookie
      name: jewelcase_session
      description: Set at login for browser clients. State-changing requests must also send `X-Jewelcase-Client`.
...
  responses:
...
    NotFound:
      description: '`not_found` — nonexistent, or out of the caller''s reach.'
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    SetupRequest:
      type: object
      required: [username, password, device]
      properties:
        username: { $ref: '#/components/schemas/Username' }
        password: { type: string, format: password }
        displayName: { type: string }
        device: { $ref: '#/components/schemas/DeviceRegistration' }
...